refactor: rename config-list to ConfigList
