+++ src/pages/index.ts
@@ -2,7 +2,7 @@ import { IncidentDetailsPage } from "./incident/IncidentDetails";
 import { IncidentListPage } from "./incident/IncidentListPage";
 import { LogsPage } from "./logs-page";
 import { CanaryPage } from "./canary";
-import { ConfigListPage } from "./config/config-list";
+import { ConfigListPage } from "./config/ConfigList";
 import { ConfigDetailsPage } from "./config/ConfigDetailsPage";
 import { ConfigChangesPage } from "./config/config-changes";
 import { ConfigDetailsChangesPage } from "./config/config-details-changes";
